Get out session info from the check, closes #22
This is a bit tricky, we supply an R user profile
(via R_PROFILE_USER) to set the `.Last` function
in the session, and this function saves the session
information to a file.

Multiple R sessions are started during an R CMD check,
and we save the information for all of them, but we
only keep one, the first one in which the checked
package is attached. This is likely the most
informative one.